bmintyApi/requirements.txt
asgiref==3.8.1
click==8.2.0
Django==5.1.6
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==25.1
djangorestframework==3.15.2
drf-yasg==1.21.7
numpy==1.26.4
pandas==2.2.0
pillow==11.1.0
psutil==6.1.1
roc-validator>=0.8.0
sqlparse==0.5.3
svgwrite==1.4.3
Tree==0.2.4
